requirements.txt
GitPython>=3.1.40
keyring>=24.3.0
pyobjc>=10.0